+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/full.php
@@ -71,7 +71,6 @@ $container->loadFromExtension('framework', [
         ],
     ],
     'property_info' => true,
-    'type_info' => true,
     'request' => [
         'formats' => [
             'csv' => [