+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/json_streamer.php
@@ -1,9 +1,6 @@
 <?php
 
 $container->loadFromExtension('framework', [
-    'type_info' => [
-        'enabled' => true,
-    ],
     'json_streamer' => [
         'enabled' => true,
     ],