+++ src/Symfony/Component/TypeInfo/Tests/DependencyInjection/Fixtures/type_info.php
@@ -1,0 +1,7 @@
+<?php
+
+$container->loadFromExtension('type_info', [
+    'aliases' => [
+        'CustomAlias' => 'int',
+    ],
+]);