+++ src/Symfony/Component/TypeInfo/Tests/DependencyInjection/Fixtures/type_info_default_config.php
@@ -1,0 +1,3 @@
+<?php
+
+$container->loadFromExtension('type_info', null);