+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -10,6 +10,8 @@ CHANGELOG
  * Register `SemaphoreBundle` automatically when the Semaphore component is installed
  * Make `framework.html_sanitizer` an alias of the `html_sanitizer` configuration provided by `HtmlSanitizerBundle`
  * Register `HtmlSanitizerBundle` automatically when the HtmlSanitizer component is installed
+ * Make `framework.type_info` an alias of the `type_info` configuration provided by `TypeInfoBundle`
+ * Register `TypeInfoBundle` automatically when the TypeInfo component is installed
  * Make `framework.workflows` an alias of the `workflow` configuration provided by `WorkflowBundle`
  * Register `WorkflowBundle` automatically when the Workflow component is installed
  * Make `framework.remote_event` an alias of the `remote_event` configuration provided by `RemoteEventBundle`